Use DOUBLE instead of DECIMAL

+++ server/migrations/postgres/00000106_create_stock_line_table/up.sql
@@ -4,8 +4,8 @@ CREATE TABLE stock_line (
     store_id TEXT NOT NULL REFERENCES store(id),
     batch TEXT,
     expiry_date TEXT,
-    cost_price_per_pack DECIMAL NOT NULL,
-    sell_price_per_pack DECIMAL NOT NULL,
+    cost_price_per_pack DOUBLE PRECISION NOT NULL,
+    sell_price_per_pack DOUBLE PRECISION NOT NULL,
     available_number_of_packs INTEGER NOT NULL,
     total_number_of_packs INTEGER NOT NULL,
     pack_size INTEGER NOT NULL
